5381: display Other filing party on edit docket entry meta; refactor into component
