+++ shared/src/business/useCases/docketEntry/updateDocketEntryMetaInteractor.js
@@ -70,6 +70,7 @@ exports.updateDocketEntryMetaInteractor = async ({
     lodged: docketEntryMeta.lodged,
     objections: docketEntryMeta.objections,
     ordinalValue: docketEntryMeta.ordinalValue,
+    otherFilingParty: docketEntryMeta.otherFilingParty,
     partyIrsPractitioner: docketEntryMeta.partyIrsPractitioner,
     partyPrimary: docketEntryMeta.partyPrimary,
     partySecondary: docketEntryMeta.partySecondary,
