+++ web-client/src/presenter/computeds/editDocketEntryMetaHelper.test.js
@@ -84,118 +84,4 @@ describe('editDocketEntryMetaHelper', () => {
       expect(result.showObjection).toBeFalsy();
     });
   });
-
-  describe('partyValidationError', () => {
-    it('should return "error" if there are party validation errors for a primary party', () => {
-      const result = runCompute(editDocketEntryMetaHelper, {
-        state: {
-          caseDetail: {
-            partyType: PARTY_TYPES.petitioner,
-          },
-          form: {
-            documentId: '123',
-          },
-          validationErrors: {
-            partyPrimary: 'error',
-          },
-        },
-      });
-      expect(result.partyValidationError).toEqual('error');
-    });
-
-    it('should return "error" if there are party validation errors for a secondary party', () => {
-      const result = runCompute(editDocketEntryMetaHelper, {
-        state: {
-          caseDetail: {
-            partyType: PARTY_TYPES.petitioner,
-          },
-          form: {
-            documentId: '123',
-          },
-          validationErrors: {
-            partySecondary: 'error',
-          },
-        },
-      });
-      expect(result.partyValidationError).toEqual('error');
-    });
-
-    it('should return "error" if there are party validation errors for a respondent party', () => {
-      const result = runCompute(editDocketEntryMetaHelper, {
-        state: {
-          caseDetail: {
-            partyType: PARTY_TYPES.petitioner,
-          },
-          form: {
-            documentId: '123',
-          },
-          validationErrors: {
-            partyIrsPractitioner: 'error',
-          },
-        },
-      });
-      expect(result.partyValidationError).toEqual('error');
-    });
-
-    it('should be falsy if there are no party validation errors at all', () => {
-      const result = runCompute(editDocketEntryMetaHelper, {
-        state: {
-          caseDetail: {
-            partyType: PARTY_TYPES.petitioner,
-          },
-          form: {
-            documentId: '123',
-          },
-          validationErrors: {},
-        },
-      });
-      expect(result.partyValidationError).toBeFalsy();
-    });
-  });
-  describe('showSecondaryParty', () => {
-    it('should show secondary party if party type is petitioner and spouse', () => {
-      const result = runCompute(editDocketEntryMetaHelper, {
-        state: {
-          caseDetail: {
-            partyType: PARTY_TYPES.petitionerSpouse,
-          },
-          form: {
-            documentId: '123',
-          },
-          validationErrors: {},
-        },
-      });
-      expect(result.showSecondaryParty).toBeTruthy();
-    });
-
-    it('should show secondary party if party type is petitioner and deceased spouse', () => {
-      const result = runCompute(editDocketEntryMetaHelper, {
-        state: {
-          caseDetail: {
-            partyType: PARTY_TYPES.petitionerDeceasedSpouse,
-          },
-          form: {
-            documentId: '123',
-          },
-          validationErrors: {},
-        },
-      });
-      expect(result.showSecondaryParty).toBeTruthy();
-    });
-
-    it('should not show secondary party if party type is anything other than petitioner and their spouse', () => {
-      const result = runCompute(editDocketEntryMetaHelper, {
-        state: {
-          caseDetail: {
-            partyType: PARTY_TYPES.petitioner,
-          },
-          form: {
-            documentId: '123',
-          },
-          validationErrors: {},
-        },
-      });
-      expect(result.showSecondaryParty).toBeFalsy();
-    });
-  });
 });
