+++ web-client/src/presenter/computeds/filingPartiesFormHelper.js
@@ -1,14 +1,20 @@
-import { find } from 'lodash';
-import { getOptionsForCategory } from './selectDocumentTypeHelper';
 import { state } from 'cerebral';
 
-export const editDocketEntryMetaHelper = (get, applicationContext) => {
-  const { eventCode } = get(state.form);
-
+export const filingPartiesFormHelper = (get, applicationContext) => {
   const caseDetail = get(state.caseDetail);
+  const validationErrors = get(state.validationErrors);
   const form = get(state.form);
 
-  const { INTERNAL_CATEGORY_MAP } = applicationContext.getConstants();
+  const {
+    INTERNAL_CATEGORY_MAP,
+    PARTY_TYPES,
+  } = applicationContext.getConstants();
+
+  const partyValidationError =
+    validationErrors &&
+    (validationErrors.partyPrimary ||
+      validationErrors.partySecondary ||
+      validationErrors.partyIrsPractitioner);
 
   const objectionDocumentTypes = [
     ...INTERNAL_CATEGORY_MAP['Motion'].map(entry => {
@@ -21,23 +27,16 @@ export const editDocketEntryMetaHelper = (get, applicationContext) => {
 
   const amendmentEventCodes = ['AMAT', 'ADMT'];
 
-  let categoryInformation;
-  find(
-    INTERNAL_CATEGORY_MAP,
-    entries => (categoryInformation = find(entries, { eventCode: eventCode })),
-  );
-
-  const optionsForCategory = getOptionsForCategory(
-    applicationContext,
-    caseDetail,
-    categoryInformation,
-  );
+  const showSecondaryParty =
+    caseDetail.partyType === PARTY_TYPES.petitionerSpouse ||
+    caseDetail.partyType === PARTY_TYPES.petitionerDeceasedSpouse;
 
   return {
-    primary: optionsForCategory,
-    showObjection:
+    noMargin:
       objectionDocumentTypes.includes(form.documentType) ||
       (amendmentEventCodes.includes(form.eventCode) &&
         objectionDocumentTypes.includes(form.previousDocument?.documentType)),
+    partyValidationError,
+    showSecondaryParty,
   };
 };
