+++ web-client/src/presenter/state.js
@@ -38,6 +38,7 @@ import { editStatisticFormHelper } from './computeds/editStatisticFormHelper';
 import { externalUserCasesHelper } from './computeds/Dashboard/externalUserCasesHelper';
 import { fileDocumentHelper } from './computeds/fileDocumentHelper';
 import { fileUploadStatusHelper } from './computeds/fileUploadStatusHelper';
+import { filingPartiesFormHelper } from './computeds/filingPartiesFormHelper';
 import {
   formattedCaseDetail,
   formattedClosedCases,
@@ -128,6 +129,7 @@ const helpers = {
   externalUserCasesHelper,
   fileDocumentHelper,
   fileUploadStatusHelper,
+  filingPartiesFormHelper,
   formattedCaseDetail,
   formattedCaseMessages,
   formattedClosedCases,
